+++ src/context/pokemonContext.js
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const PokemonContext = React.createContext(null)
