+++ src/routes/Game/Finish/Finish.js
@@ -1,0 +1,7 @@
+const Finish = () => {
+  return(
+    <h1>Game Over</h1>
+  )
+}
+
+export default Finish
